feat: add base class for Contact related ViewModels
